Move variable and lists storage to `Target` instead of `Clone`
